# backend/internal/interfaces/httpapi/openapi.yaml
openapi: 3.0.3
info:
  title: Calculator API
  version: 1.0.0
  description: Stateless calculator API with domain-level validation.
servers:
  - url: http://localhost:8080
    description: Local backend
tags:
  - name: Calculations
  - name: Health
paths:
  /api/v1/calculations:
    post:
      tags: [Calculations]
      summary: Calculate a result
      operationId: calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculationRequest'
            examples:
              addition:
                value: { operation: add, a: 12, b: 8 }
              squareRoot:
                value: { operation: sqrt, a: 81 }
              percentage:
                value: { operation: percentage, a: 25, b: 200 }
      responses:
        '200':
          description: Calculation completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResponse'
        '400':
          description: Malformed, incomplete, or invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Mathematically undefined or non-finite calculation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /healthz:
    get:
      tags: [Health]
      summary: Check API health
      operationId: healthCheck
      responses:
        '200':
          description: API is healthy.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status:
                    type: string
                    example: ok
components:
  schemas:
    CalculationRequest:
      type: object
      additionalProperties: false
      required: [operation, a]
      properties:
        operation:
          type: string
          enum: [add, subtract, multiply, divide, power, sqrt, percentage]
          description: Operation to perform. The `b` operand is required except for `sqrt`.
        a:
          type: number
          format: double
          description: First operand, which must be finite.
        b:
          type: number
          format: double
          description: Second operand. For percentage, this is the base value.
    CalculationResponse:
      type: object
      required: [operation, result]
      properties:
        operation:
          type: string
          example: add
        result:
          type: number
          format: double
          example: 20
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: calculation_error
            message:
              type: string
              example: division by zero